requirements.txt
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
open-clip-torch>=2.24.0
streamlit>=1.30.0
Pillow>=10.0.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
safetensors>=0.4.0
